requirements.txt
openai
flask
joblib
pytest
python-dotenv
